Created sort hook
